Move embed tests into separate file. Delete test_views file.
